practice_file/0206/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-mail>=1.6.1",
    "fastapi[standard]>=0.128.2",
    "kafka-python>=2.3.0",
    "pydantic-settings>=2.12.0",
]
